+++ frontend/src/app/on-time/controls/controls.component.ts
@@ -29,9 +29,9 @@ import {
   Period,
   Preset,
 } from "../../shared/components/date-range/date-range.types";
-import { PanelService } from "../../shared/components/panel/panel.service";
 import { ifNullOrUndefinedReturnEmptyString } from "../../shared/rxjs-operators";
-import { ControlsComponent } from "../../shared/components/controls/controls.component";
+import { OnTimePanelService } from "./on-time-panel.service";
+import { PanelService } from "../../shared/components/panel/panel.service";
 
 export type TimingPoints = "all-stops" | "timing-points";
 
@@ -40,8 +40,7 @@ export type TimingPoints = "all-stops" | "timing-points";
   templateUrl: "controls.component.html",
   styleUrls: ["./controls.component.scss"],
 })
-export class OnTimeControlsComponent
-  extends ControlsComponent
+export class ControlsComponent
   implements OnInit, OnChanges, AfterViewInit, OnDestroy
 {
   @Input() showAdminAreas = true;
@@ -50,30 +49,151 @@ export class OnTimeControlsComponent
 
   filtersSubject = new BehaviorSubject<PerformanceFiltersInputType>({});
 
+  dateRange = new FormControl(
+    this.getDateTimeParams(this.route.snapshot.queryParamMap),
+    {
+      nonNullable: true,
+    },
+  );
+
   onDestroy$ = new Subject<void>();
 
   @ViewChild(FiltersComponent) filtersComponent?: FiltersComponent;
 
+  // TODO ABOD-350 prefer a form to custom binding
+  get timingPointsOption(): TimingPoints {
+    return this.filtersSubject.value.timingPointsOnly
+      ? "timing-points"
+      : "all-stops";
+  }
+
+  set timingPointsOption(timingPointsOption: TimingPoints) {
+    const allStops = timingPointsOption === "all-stops" || null;
+    this.router.navigate([], {
+      queryParams: { allStops, timingPointsOnly: null },
+      queryParamsHandling: "merge",
+    });
+  }
+
+  get matchType(): MatchType {
+    return this.filtersSubject.value.matchType ?? MatchType.Evidenced;
+  }
+
+  set matchType(match_type: MatchType) {
+    this.router.navigate([], {
+      queryParams: { match_type },
+      queryParamsHandling: "merge",
+    });
+  }
+
+  /** @deprecated this will be removed in ABOD-350 */
+  get filters(): PerformanceFiltersInputType {
+    return this.filtersSubject.value;
+  }
+
   constructor(
-    protected router: Router,
-    protected route: ActivatedRoute,
+    private router: Router,
+    private route: ActivatedRoute,
+    private dateRangeService: DateRangeService,
     public panelService: PanelService,
-    protected dateRangeService: DateRangeService,
-  ) {
-    super(router, route, dateRangeService);
-    this.router = router;
-    this.panelService = panelService;
+    public onTimePanel: OnTimePanelService,
+  ) {}
+
+  getDateTimeParams(queryParams: ParamMap): FromToPreset {
+    let from, to: DateTime;
+    let preset: Preset;
+
+    if (queryParams.get("from") && queryParams.get("to")) {
+      from = DateTime.fromFormat(
+        queryParams.get("from")!,
+        "yyyy-MM-dd",
+      ).toLocal();
+      to = DateTime.fromFormat(queryParams.get("to")!, "yyyy-MM-dd")
+        .toLocal()
+        .plus({ days: 1 }); // date range is exclusive on the to date
+      preset = Preset.Custom;
+    } else {
+      preset = (queryParams.get("preset") as Period) ?? Period.Last7;
+      const presetRange = this.dateRangeService.calculatePresetPeriod(
+        preset,
+        DateTime.local(),
+      );
+      from = presetRange.from;
+      to = presetRange.to;
+    }
+
+    return { from, to, preset };
   }
 
   getPerformanceFilters(
     paramMap: ParamMap,
     queryParams: ParamMap,
   ): PerformanceFiltersInputType {
-    const filters: PerformanceFiltersInputType = {};
+    const filters: PerformanceFiltersInputType = {
+      timingPointsOnly: true,
+      matchType: MatchType.Evidenced,
+    };
 
     if (paramMap.get("nocCode")) {
       const operatorId = ifNullOrUndefinedReturnEmptyString(this.operatorId);
       filters.operatorIds = [operatorId];
+    }
+    const lineId = paramMap.get("lineId");
+    if (lineId) {
+      filters.lineIds = [lineId];
+    }
+
+    if (queryParams.get("dayOfWeek")) {
+      const dayOfWeekFlags: DayOfWeekFlagsInputType = {
+        monday: false,
+        tuesday: false,
+        wednesday: false,
+        thursday: false,
+        friday: false,
+        saturday: false,
+        sunday: false,
+      };
+
+      queryParams
+        .get("dayOfWeek")
+        ?.split(",")
+        .map((day) => {
+          if (day as keyof DayOfWeekFlagsInputType)
+            dayOfWeekFlags[day as keyof DayOfWeekFlagsInputType] = true;
+        });
+
+      filters.dayOfWeekFlags = dayOfWeekFlags;
+    }
+
+    if (queryParams.get("startTime")) {
+      filters.startTime = queryParams.get("startTime");
+    }
+
+    if (queryParams.get("endTime")) {
+      filters.endTime = queryParams.get("endTime");
+    }
+
+    if (queryParams.get("minDelay")) {
+      filters.minDelay = parseInt(queryParams.get("minDelay")!);
+    }
+
+    if (queryParams.get("maxDelay")) {
+      filters.maxDelay = parseInt(queryParams.get("maxDelay")!);
+    }
+
+    if (queryParams.has("timingPointsOnly")) {
+      filters.timingPointsOnly =
+        queryParams.get("timingPointsOnly") === "true" || undefined;
+    }
+
+    if (queryParams.has("match_type")) {
+      filters.matchType =
+        (queryParams.get("match_type") as MatchType) || MatchType.Evidenced;
+    }
+
+    if (queryParams.has("allStops")) {
+      filters.timingPointsOnly =
+        queryParams.get("allStops") !== "true" || undefined;
     }
 
     if (queryParams.has("adminAreaId")) {
@@ -194,7 +314,7 @@ export class OnTimeControlsComponent
   }
 
   resetFilters() {
-    this.filtersComponent?.resetFilters();
+    this.onTimePanel.resetFilters();
   }
 
   overviewModeChanged(overview: string) {
@@ -210,7 +330,7 @@ export class OnTimeControlsComponent
 
   setFilterPanelComponent() {
     this.panelService.setComponent({
-      component: FiltersComponent,
+      component: this.onTimePanel.getComponent(),
       inputs: [
         {
           name: "filters",
